requirements.txt
# OECS static site scripts
# Use with a virtual environment: python -m venv .venv && source .venv/bin/activate (or .venv\Scripts\activate on Windows)

requests>=2.28.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pycryptodome>=3.18.0
markdownify>=0.11.0
PyYAML>=6.0
